update  Diagram.md
Each entity represents a fictional character, and all of these characters have some form of relationship with each other. These entities are depicted in the class diagram, with the contents of each class outlining both positive and negative traits. These details help provide a brief overview of the characters' attributes. Each relationship between the entities is represented with specific keywords and symbols to describe the nature of the relationship. For example, Hinx has a strong relationship with Damian (depicted as *--), who is also in love with Hinx, as represented in the diagram.